+++ lib/jarbler/builder.rb
@@ -95,11 +95,15 @@ module Jarbler
     # @return [void]
     def copy_needed_gems_to_staging(staging_dir, ruby_minor_version)
       gem_target_location = "#{staging_dir}/gems/jruby/#{ruby_minor_version}"
+
+      # Replace universal-java-XX with the correct platform for the current Java version after unzipping of the jar file
+      extension_target_location = "#{gem_target_location}/extensions/universal-java-XX/#{ruby_minor_version}"
+
       FileUtils.mkdir_p("#{gem_target_location}/bin")
       FileUtils.mkdir_p("#{gem_target_location}/build_info")
       FileUtils.mkdir_p("#{gem_target_location}/cache")
       FileUtils.mkdir_p("#{gem_target_location}/doc")
-      FileUtils.mkdir_p("#{gem_target_location}/extensions")
+      FileUtils.mkdir_p(extension_target_location)
       FileUtils.mkdir_p("#{gem_target_location}/gems")
       FileUtils.mkdir_p("#{gem_target_location}/specifications")
       FileUtils.mkdir_p("#{gem_target_location}/bundler/bin")
@@ -115,8 +119,10 @@ module Jarbler
         # differentiate between Gems from git/bundler and Gems from rubygems
         if spec.source.is_a?(Bundler::Source::Git)
           # Copy the Gem from bundler/gems including the gemspec
+          debug "Adding Bundler Gem from dir '#{spec.gem_dir}' into jar file at temporary location '#{gem_target_location}/gems'"
           file_utils_copy(spec.gem_dir, "#{gem_target_location}/bundler/gems")
           spec.executables.each do |executable|
+            debug "Adding executable of Bundler Gem from dir '#{spec.bin_dir}#{executable}/' into jar file at temporary location '#{gem_target_location}/bundler/bin'"
             file_utils_copy("#{spec.bin_dir}/#{executable}", "#{gem_target_location}/bundler/bin")
           end
         else  # Gem is from rubygems
@@ -124,10 +130,19 @@ module Jarbler
           #       Should the default gems are also copied to the staging directory?
           unless spec.default_gem?  # Do not copy default gems, because they are already included in the jruby jars standard library
             # copy the Gem and gemspec separately
+            debug "Adding local Gem from dir '#{spec.gem_dir}' into jar file at temporary location '#{gem_target_location}/gems'"
             file_utils_copy(spec.gem_dir, "#{gem_target_location}/gems")
+
+            if spec.extension_dir && Dir.exist?(spec.extension_dir)
+              debug "Adding extension from dir '#{spec.extension_dir}' into jar file at temporary location '#{extension_target_location}'"
+              puts "Adding extension from dir '#{spec.extension_dir}' into jar file but extension is not for platform 'universal-java-xx'!!!" unless spec.extension_dir['universal-java']
+              file_utils_copy(spec.extension_dir, extension_target_location)
+            end
+
             # spec.loaded_from contains the path to the gemspec file including the path prefix "default/" for default gems
             file_utils_copy(spec.loaded_from, "#{gem_target_location}/specifications")
             spec.executables.each do |executable|
+              debug "Adding executable of local Gem from dir '#{spec.bin_dir}#{executable}/' into jar file at temporary location '#{gem_target_location}/bin'"
               file_utils_copy("#{spec.bin_dir}/#{executable}", "#{gem_target_location}/bin")
             end
           end